Create Led blinking with arduino uno
We will interface an LED (light-emitting diode) to the Arduino UNO board. An LED is a simple diode that emits light in a forward bias. We will write an LED-blinking program on the Arduino IDE and download it to the microcontroller board. The program simply turns ON and OFF LED with some delay between them.

Table of Content

Blinking an LED
Working procedure
Components Required
Arduino Code
Deployment Using Arduino IDE
Circuit Diagram
Applications